Move slices pkg to std and rename it to hslice
