--- pkg/slices/int64_test.go
@@ -1,1 +1,0 @@
-package slices
